move all reduxConstants into one file:
