+++ src/js/actions/DirectionActions.js
@@ -1,4 +1,4 @@
-import { UPDATE_DIRECTION} from './types';
+import { UPDATE_DIRECTION} from '../reduxConstants';
 
 export const updateDirection = (updateDirection) => {          // updates left to right direction for countries that read in that language
