--- src/js/actions/types.js
@@ -1,15 +1,0 @@
-export const FINISHED_PLAYING = 'finished_playing';
-export const PLAY_AUDIO = 'play_audio';
-export const STOP_AUDIO = 'stop_audio';
-export const UPDATE_PLAYLIST = 'update_playlist';
-export const MULTIPLE_TAKES = 'multiple_takes';
-export const UPDATE_MODE = 'update_mode';   // update mode in the object that contains all the data of the take ( chunk or verse);
-export const UPDATE_TIME = 'update_time';   // update time displayed in the audioplayer
-export const UPDATE_AUDIO_LENGTH = 'update_audio_length';
-export const PLAY_TAKE = 'play_take';
-export const CLEAR_PLAYLIST = 'clear_playlist';
-export const REMOVE_TAKE_FROM_PLAYLIST = 'remove_take_from_playlist';
-export const RESET_AUDIOPLAYER = 'remove_audioplayer';
-export const TAKE_ID = 'take_id';
-export const UPDATE_LANGUAGE = 'update_language';
-export const UPDATE_DIRECTION = 'update_direction';
