--- src/js/reducers/types.js
@@ -1,11 +1,0 @@
-export const PLAY_PAUSE_SELECTED_TAKE='PLAY_PAUSE_SELECTED_TAKE';
-export const PAUSE_SELECTED_TAKE='PAUSE_SELECTED_TAKE';
-export const LOADING='LOADING';
-export const FETCH_ALTERNATE_TAKES='FETCH_ALTERNATE_TAKES';
-export const FETCH_CHUNKS='FETCH_CHUNKS';
-export const UPDATE_ACTIVE_CHUNK_INDEX = 'UPDATE_ACTIVE_CHUNK_INDEX';
-export const SWAP_TAKE = 'SWAP_TAKE';
-export const UNDO_SWAP = 'UNDO_SWAP';
-export const UPDATE_ALTERNATE_TAKES = 'UPDATE_ALTERNATE_TAKES';
-export const FINISH_PLAYING = 'FINISH_PLAYING';
-export const CLEAR_ALTERNATE_TAKES = 'CLEAR_ALTERNATE_TAKES';
